--- scripts/shared/twofactor/__init__.py
@@ -1,5 +1,0 @@
-from scripts.shared.twofactor.auth_code import get_secure_code
-from scripts.shared.twofactor.code_handler import CodeHandler
-from scripts.shared.twofactor.main import send_code_over_2f
-
-__all__ = ["get_secure_code", "send_code_over_2f", "CodeHandler"]
